add BaseCartAndBookmarkWrapper
