+++ src/test/java/wrappers/BookmarksPageWrapper.java
@@ -1,0 +1,11 @@
+package wrappers;
+
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.WebElement;
+
+public class BookmarksPageWrapper extends BaseCartAndBookmarkWrapper {
+
+    public BookmarksPageWrapper(WebDriver driver, WebElement cardCart) {
+        super(driver, cardCart);
+    }
+}
